maint: move dict-id-replacer to utils
